Start splitting kubeval up into separate packages
First pass at splitting the monolithic package into separate packages.
This also facilitates the use of kubeval as a library by other tools
which has been discussed a few times.

This is the first pass at this, I want to split the Validate function
down some, and make sure everything is documented. I'll get this merged
quickly to avoid any pain with PRs needing rebasing.
