+++ .gitignore
@@ -1,4 +1,3 @@
-kubeval
 vendor
 bin
 releases
