+++ kubeval/utils.go
@@ -1,0 +1,22 @@
+package kubeval
+
+// Based on https://stackoverflow.com/questions/40737122/convert-yaml-to-json-without-struct-golang
+// We unmarshal yaml into a value of type interface{},
+// go through the result recursively, and convert each encountered
+// map[interface{}]interface{} to a map[string]interface{} value
+// required to marshall to JSON.
+func convertToStringKeys(i interface{}) interface{} {
+	switch x := i.(type) {
+	case map[interface{}]interface{}:
+		m2 := map[string]interface{}{}
+		for k, v := range x {
+			m2[k.(string)] = convertToStringKeys(v)
+		}
+		return m2
+	case []interface{}:
+		for i, v := range x {
+			x[i] = convertToStringKeys(v)
+		}
+	}
+	return i
+}
